Add files via upload
Portal Master is a teleporter NPC for 3.3.5a.
It has 130+ teleports including all zones in wow.
Portal Master was created with blizzlike feeling and usability in mind.
The teleporter has been in use on many servers for years.
Unlike many other teleporters it does not use weird colors or icons in menus.
The teleports are categorized and you can only see your own faction teleports.
All teleports have the zone’s level requirement. This means that you can not teleport anywhere when at a lower level.
The teleporter is spawned by default in every main city at the location of teleportation.
The download options include a TrinityCore version, a MaNGOS version as well as a teleport option creating SQL for TrinityCore.
Deleting the default installation can be done with the deletion SQL included with the installation SQL.
The teleporter is completely created with SQL which means you can install and edit it without recompiling your core.